+++ src/plugin/gui/gui_button_widgets.hpp
@@ -110,6 +110,32 @@ PUBLIC Style IconButton(imgui::Context const& imgui) {
     return s;
 }
 
+static Style MuteSoloButton(imgui::Context const& imgui, bool is_solo) {
+    Style s {};
+    s.type = LayoutAndSizeType::IconOrText;
+    s.main_cols.reg = LiveCol(imgui, UiColMap::MuteSoloButtonTextRegular);
+    s.main_cols.on = LiveCol(imgui, UiColMap::MuteSoloButtonTextOn);
+    s.main_cols.hot_on = LiveCol(imgui, UiColMap::MuteSoloButtonTextOnHover);
+    s.main_cols.hot_off = LiveCol(imgui, UiColMap::MuteSoloButtonTextHover);
+    s.main_cols.active_on = s.main_cols.hot_on;
+    s.main_cols.active_off = s.main_cols.hot_off;
+    s.back_cols.reg = {};
+    s.back_cols.on = LiveCol(imgui, is_solo ? UiColMap::SoloButtonBackOn : UiColMap::MuteButtonBackOn);
+    s.back_cols.hot_on = s.back_cols.on;
+    s.back_cols.hot_off = s.back_cols.reg;
+    s.back_cols.active_on = s.back_cols.on;
+    s.back_cols.active_off = s.back_cols.reg;
+    s.icon_or_text.add_margin_x = false;
+    s.icon_or_text.justification = TextJustification::Centred;
+    s.icon_scaling = Style::k_regular_icon_scaling;
+    s.corner_rounding_flags = is_solo ? 0b0110 : 0b1001;
+    return s;
+}
+
+PUBLIC Style MuteButton(imgui::Context const& imgui) { return MuteSoloButton(imgui, false); }
+
+PUBLIC Style SoloButton(imgui::Context const& imgui) { return MuteSoloButton(imgui, true); }
+
 PUBLIC Style SettingsWindowButton(imgui::Context const& imgui) {
     auto s = IconButton(imgui);
     s.type = LayoutAndSizeType::IconAndText;